+++ lib/oneview-sdk/resource/api1200/c7000/logical_enclosure.rb
@@ -16,16 +16,6 @@ module OneviewSDK
     module C7000
       # Logical Enclosure resource implementation on API1200 C7000
       class LogicalEnclosure < OneviewSDK::API1000::C7000::LogicalEnclosure
-        # Create a resource object, associate it with a client, and set its properties.
-        # @param [OneviewSDK::Client] client The client object for the OneView appliance
-        # @param [Hash] params The options for this resource (key-value pairs)
-        # @param [Integer] api_ver The api version to use when interracting with this resource.
-        def initialize(client, params = {}, api_ver = nil)
-          @data ||= {}
-          # Default values:
-          @data['type'] ||= 'LogicalEnclosureV5'
-          super
-        end
       end
     end
   end